requirements.txt
pyyaml
aiohttp
aiokafka
asyncpg
